Loss of precision when converting from double to int and double to float in FrameView::zoomAnimatorTransformChanged() https://bugs.webkit.org/show_bug.cgi?id=69739
Source/WebCore:

Reviewed by Simon Fraser.

Change 'double' parameters in FrameView:: & ScrollableArea::zoomAnimatorTransformChanged()
to 'float' to avoid loss of precision warnings when invoking setPageScalefactor.

Test coverage provided by existing zoom-animator tests.

* page/FrameView.cpp:
(WebCore::FrameView::zoomAnimatorTransformChanged):
* page/FrameView.h:
* platform/ScrollableArea.cpp:
(WebCore::ScrollableArea::zoomAnimatorTransformChanged):
* platform/ScrollableArea.h:

Source/WebKit2:

Change 'double' parameters to 'float' to match changes in FrameView.

Reviewed by Simon Fraser.

* WebProcess/Plugins/PDF/BuiltInPDFView.h:
(WebKit::BuiltInPDFView::zoomAnimatorTransformChanged):


Canonical link: https://commits.webkit.org/85921@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@97188 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ Source/WebCore/page/FrameView.h
@@ -170,7 +170,7 @@ public:
     virtual void repaintFixedElementsAfterScrolling();
     virtual bool shouldRubberBandInDirection(ScrollDirection) const;
 
-    virtual void zoomAnimatorTransformChanged(double, double, double, ZoomAnimationState);
+    virtual void zoomAnimatorTransformChanged(float, float, float, ZoomAnimationState);
 
     String mediaType() const;
     void setMediaType(const String&);
